feat(containers): split components
separate components into show components and container components
