+++ src/router/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Routes, Route } from 'react-router-dom'
 import LoginPage from '../components/Login'
-import Home from '../components/Home'
+import Home from '../containers/Home'
 import { ComContainer } from './style'
 
 export default function routes() {
